Added auth for nodejs rest api with protected routes

+++ src/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goals = require("../models/goalmodels");
+const User = require("../models/usermodels");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goals.find();
+  const goals = await Goals.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -12,7 +13,10 @@ const postGoals = asyncHandler(async (req, res) => {
     throw new Error("Goals field is required");
   }
 
-  const goal = await Goals.create({ text: req.body.text });
+  const goal = await Goals.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
   res.status(201).json(goal);
 });
 
@@ -21,6 +25,22 @@ const putGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
+  }
+
+  if (!req.body.text) {
+    res.status(400);
+    throw new Error("Text field is required");
+  }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goals.findByIdAndUpdate(req.params.id, req.body, {
@@ -34,6 +54,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404).json({ error: "Goal not Found" });
   }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   await Goals.deleteOne({ _id: req.params.id });
   res
     .status(200)
